// src/types/pokemon.ts
export interface PokemonAbility {
  ability: {
    name: string;
    url: string;
  };
  is_hidden: boolean;
  slot: number;
}

export interface PokemonType {
  slot: number;
  type: {
    name: string;
    url: string;
  };
}

export interface PokemonSprites {
  front_default: string;
  front_shiny: string | null;
  other?: {
    'official-artwork'?: {
      front_default: string;
      front_shiny: string | null;
    };
  };
}

export interface PokemonDetails {
  id: number;
  name: string;
  height: number;
  abilities: PokemonAbility[];
  types: PokemonType[];
  sprites: PokemonSprites;
}

export interface PokemonListItem {
  name: string;
  url: string;
}

export interface PokemonListResponse {
  count: number;
  next: string | null;
  previous: string | null;
  results: PokemonListItem[];
}